// examen-front/src/app/model/clienteModelo.ts
export class ClienteModel{
    id_cliente: number= 0;
    nombre: string = '';
    apellido: string = '';
    numero_identificacion: string = '';
    fecha_nacimiento: string ='';
    direccion: string ='';
    correo_electronico: string = '';
    telefono: string = '';
    estado: boolean = false;
}